Moved postgrs configuration to shared
