+++ Modules/AlledrogO.Post/AlledrogO.Post.Api/Extensions.cs
@@ -11,7 +11,7 @@ public static class Extensions
     public static IServiceCollection AddPostModule(this IServiceCollection services,  IConfiguration configuration)
     {
         services.AddApplication();
-        services.AddInfrastructure(configuration);   
+        services.AddInfrastructure();   
         return services;
     }
