+++ Modules/AlledrogO.Post/AlledrogO.Post.Infrastructure/AlledrogO.Post.Infrastructure.csproj
@@ -13,7 +13,6 @@
     <ItemGroup>
       <PackageReference Include="Microsoft.EntityFrameworkCore" Version="8.0.4" />
       <PackageReference Include="Microsoft.EntityFrameworkCore.Relational" Version="8.0.4" />
-      <PackageReference Include="Npgsql.EntityFrameworkCore.PostgreSQL" Version="8.0.2" />
     </ItemGroup>
 
     <ItemGroup>
