+++ Modules/AlledrogO.Post/AlledrogO.Post.Infrastructure/EF/Extensions.cs
@@ -1,7 +1,7 @@
 using AlledrogO.Post.Application.Contracts;
 using AlledrogO.Post.Infrastructure.EF.Contexts;
-using AlledrogO.Post.Infrastructure.EF.Options;
 using AlledrogO.Post.Infrastructure.EF.Repositories;
+using AlledrogO.Shared.Database;
 using AlledrogO.Shared.Options;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Configuration;
@@ -11,20 +11,17 @@ namespace AlledrogO.Post.Infrastructure.EF;
 
 internal static class Extensions
 {
-    public static IServiceCollection AddPostgres(this IServiceCollection services, IConfiguration configuration)
+    public static IServiceCollection AddDatabase(this IServiceCollection services)
     {
         services.AddScoped<IPostRepository, PostgresPostRepository>();
         services.AddScoped<IAuthorRepository, PostgresAuthorRepository>();
         services.AddScoped<ITagRepository, PostgresTagRepository>();
-
-        var options = configuration.GetOptions<PostgresOptions>("Postgres")
-            ?? throw new Exception("Postgres options not found");
         
-        services.AddDbContext<WriteDbContext>(ctx =>
-            ctx.UseNpgsql(options.ConnectionString));
-        services.AddDbContext<ReadDbContext>(ctx =>
-            ctx.UseNpgsql(options.ConnectionString));
+        services.AddPostgres<WriteDbContext>();
+        services.AddPostgres<ReadDbContext>();
         return services;
     }
+    
+    
 
 }
