--- Modules/AlledrogO.Post/AlledrogO.Post.Infrastructure/EF/Options/PostgresOptions.cs
@@ -1,6 +1,0 @@
-namespace AlledrogO.Post.Infrastructure.EF.Options;
-
-public class PostgresOptions
-{
-    public string ConnectionString { get; set; }
-}
