+++ Modules/AlledrogO.Post/AlledrogO.Post.Infrastructure/Extensions.cs
@@ -2,6 +2,7 @@ using AlledrogO.Post.Infrastructure.EF;
 using AlledrogO.Post.Infrastructure.Logging;
 using AlledrogO.Shared;
 using AlledrogO.Shared.Commands;
+using AlledrogO.Shared.Queries;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 
@@ -9,9 +10,9 @@ namespace AlledrogO.Post.Infrastructure;
 
 public static class Extensions
 {
-    public static IServiceCollection AddInfrastructure(this IServiceCollection services, IConfiguration configuration)
+    public static IServiceCollection AddInfrastructure(this IServiceCollection services)
     {
-        services.AddPostgres(configuration);
+        services.AddDatabase();
         services.AddQueries();
         
         services.TryDecorate(typeof(ICommandHandler<>), typeof(LoggingCommandHandlerDecorator<>));
