+++ Shared/AlledrogO.Shared/Commands/Extensions.cs
@@ -1,0 +1,24 @@
+using System.Reflection;
+using Microsoft.Extensions.DependencyInjection;
+
+namespace AlledrogO.Shared.Commands;
+
+public static class Extensions
+{
+    public static IServiceCollection AddCommands(this IServiceCollection services)
+    {
+        var assembly = Assembly.GetCallingAssembly();
+       
+        services.AddSingleton<ICommandDispatcher, InMemoryCommandDispatcher>();
+        services.Scan(s => s.FromAssemblies(assembly)
+            .AddClasses(c => c.AssignableTo(typeof(ICommandHandler<>)))
+            .AsImplementedInterfaces()
+            .WithScopedLifetime());
+       
+        services.Scan(s => s.FromAssemblies(assembly)
+            .AddClasses(c => c.AssignableTo(typeof(ICommandHandler<,>)))
+            .AsImplementedInterfaces()
+            .WithScopedLifetime());
+        return services;
+    } 
+}
