+++ Shared/AlledrogO.Shared/Database/PostgresOptions.cs
@@ -1,0 +1,6 @@
+namespace AlledrogO.Shared.Database;
+
+internal sealed class PostgresOptions
+{
+    public string ConnectionString { get; set; }
+}
