+++ Shared/AlledrogO.Shared/Queries/Extensions.cs
@@ -1,0 +1,19 @@
+using System.Reflection;
+using Microsoft.Extensions.DependencyInjection;
+
+namespace AlledrogO.Shared.Queries;
+
+public static class Extensions
+{
+    public static IServiceCollection AddQueries(this IServiceCollection services)
+    {
+        var assembly = Assembly.GetCallingAssembly();
+       
+        services.AddSingleton<IQueryDispatcher, InMemoryQueryDispatcher>();
+        services.Scan(s => s.FromAssemblies(assembly)
+            .AddClasses(c => c.AssignableTo(typeof(IQueryHandler<,>)))
+            .AsImplementedInterfaces()
+            .WithScopedLifetime());
+        return services;
+    }
+}
